Refactor CLI connect prompting and tests
