+++ gestalt/cli/src/lib.rs
@@ -4,5 +4,6 @@ pub mod cli;
 pub mod commands;
 pub mod config;
 pub mod credentials;
+pub mod interactive;
 pub mod output;
 pub mod params;
